refactor(provider-node): extract Subxt clients to separate files for coverage exclusion
Move production Subxt chain client implementations into dedicated *_subxt.rs
files so they can be excluded from CI coverage metrics alongside command.rs,
cli.rs, and main.rs. This shrinks the coverage denominator to only testable
code, giving a more accurate picture of actual test coverage.
