+++ .github/workflows/check.yml
@@ -191,7 +191,7 @@ jobs:
 
           PROVIDER_COV=$(cargo llvm-cov $TEST_FLAGS \
             -p storage-provider-node \
-            --ignore-filename-regex='(\.cargo|rustc)' \
+            --ignore-filename-regex='(\.cargo|rustc|_subxt\.rs|command\.rs|cli\.rs|main\.rs)' \
             --summary-only 2>&1 \
             | grep '^TOTAL' | awk '{print $10}' | tr -d '%')
           echo "PR_PROVIDER_COV=$PROVIDER_COV" >> $GITHUB_ENV
@@ -228,7 +228,7 @@ jobs:
           if [ -n "$TEST_FLAGS" ]; then
             BASE_PROVIDER_COV=$(cargo llvm-cov $TEST_FLAGS \
               -p storage-provider-node \
-              --ignore-filename-regex='(\.cargo|rustc)' \
+              --ignore-filename-regex='(\.cargo|rustc|_subxt\.rs|command\.rs|cli\.rs|main\.rs)' \
               --summary-only 2>&1 \
               | grep '^TOTAL' | awk '{print $10}' | tr -d '%')
           else
