+++ provider-node/src/lib.rs
@@ -9,10 +9,13 @@
 //! - Coordinating provider-initiated checkpoints
 
 pub mod agreement_coordinator;
+pub mod agreement_coordinator_subxt;
 pub mod api;
 pub mod auth;
 pub mod challenge_responder;
+pub mod challenge_responder_subxt;
 pub mod checkpoint_coordinator;
+pub mod checkpoint_coordinator_subxt;
 pub mod cli;
 pub mod command;
 pub mod error;
@@ -21,6 +24,7 @@ pub mod fs_index;
 pub mod mmr;
 pub mod replica_sync;
 pub mod replica_sync_coordinator;
+pub mod replica_sync_coordinator_subxt;
 pub mod s3_api;
 pub mod s3_index;
 pub mod storage;
